Parte 2 do trabalho, serviços utilizando stream

+++ src/main/java/br/com/trabalho1/mateus/dto/output/ProdutoDtoOutput.java
@@ -15,7 +15,7 @@ public class ProdutoDtoOutput {
 
     private String descricao;
 
-    private Long quantidadeEstoque;
+    private Integer quantidadeEstoque;
 
     private BigDecimal precoVenda;
